+++ src/com/urise/webapp/storage/ArrayStorage.java
@@ -24,7 +24,7 @@ public class ArrayStorage extends AbstractArrayStorage {
 
     @Override
     protected void doDelete(Integer index) {
-        storage[(int) index] = storage[size - 1];
+        storage[index] = storage[size - 1];
         storage[size - 1] = null;
         size--;
     }
